task are no longer allowed to request mode changes to modes they are currently in. Was originally just a programmer 'dont be dumb' thing, but it is posible that it could be done accidentally and the entire system should not waste time rescheduling

+++ main_binaries/task_manager.cpp
@@ -169,6 +169,9 @@ void sigrt1_handler(int signum){
 
 void modify_self(int new_mode){
 
+	if (new_mode == schedule.get_task(task_index)->get_current_mode())
+		return;
+
 	schedule.get_task(task_index)->set_current_mode(new_mode, true);
 	killpg(process_group, SIGRTMIN+0);
